third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-tabbed-editor-opens-filesystem-uisourcecode.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../debugger-test.js"></script>
<script src="../isolated-filesystem-test.js"></script>
<script src="./persistence-test.js"></script>
<script src="./resources/foo.js"></script>
<script>

function test() {
  var testMapping = BindingsTestRunner.initializeTestMapping();
  var fs = new BindingsTestRunner.TestFileSystem('file:///var/www');
  var fsEntry = BindingsTestRunner.addFooJSFile(fs);
  fs.reportCreated(function() {});
  testMapping.addBinding('foo.js');
  BindingsTestRunner.waitForBinding('foo.js').then(onBindingCreated);

  function onBindingCreated(binding) {
    TestRunner.addResult('Binding created: ' + binding);
    dumpEditorTabs('Opened tabs before opening any UISourceCodes:');
    TestRunner.addResult('request open uiSourceCode: ' + binding.fileSystem.url());
    UI.panels.sources.showUISourceCode(binding.network, 0, 0);
    dumpEditorTabs('Opened tabs after opening UISourceCode:');
    TestRunner.completeTest();
  }

  function dumpEditorTabs(title) {
    var editorContainer = UI.panels.sources._sourcesView._editorContainer;
    var openedUISourceCodes = editorContainer._tabIds.keysArray();
    openedUISourceCodes.sort((a, b) => a.url().compareTo(b.url()));
    TestRunner.addResult(title);
    for (code of openedUISourceCodes)
      TestRunner.addResult('    ' + code.url());
  }
};
</script>
</head>
<body onload="runTest()">
<p>Verify that for a fileSystem UISourceCode with persistence binding TabbedEditorContainer opens filesystem UISourceCode.</p>
</body>
</html>
