third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-tabbed-editor-opens-filesystem-uisourcecode-expected.txt

Verify that for a fileSystem UISourceCode with persistence binding TabbedEditorContainer opens filesystem UISourceCode.

Binding created: {
       network: http://127.0.0.1:8000/inspector/persistence/resources/foo.js
    fileSystem: file:///var/www/inspector/persistence/resources/foo.js
    exactMatch: false
}
Opened tabs before opening any UISourceCodes:
request open uiSourceCode: file:///var/www/inspector/persistence/resources/foo.js
Opened tabs after opening UISourceCode:
    file:///var/www/inspector/persistence/resources/foo.js
